Ability to map org-mode keywords to JIRA status.
Allow to look for a custom set alist of jira status and known org mode
keywords and use that as the inserted todo item status.

This preserves the user preference to first use the jira status as the
todo state, next it will look for the aforementioned alist if
existing, finally check if the jira status corresponds to a known
"DONE" state, else returning the default "TODO".