Allow to disable comments downloading during issue rendering
This change with org-jira-download-comments set to nil aims to be
something in between org-jira-get-issues-headonly call and
org-jira-get-issues.

Some users may not find downloading issues comments to be very useful, as
they don't use org-jira package for reading comments. In that case
rendering comments could bring too much delay to synchronizations process
and increase size of the org-mode file without significant reason. This
mostly problem with huge issue lists.

Although these is org-jira-get-issues-headonly for rendering only
headlines of the issues, it could be not enough, as this renderer don't
download anything except headline itself. In other words various issue
properties are not downloaded.