fix: simplify arguments passed to `completing-read`
It's possible that the arguments that are expected by `completing-read`
have changed in recent versions of Emacs - the original argument order
as passed doesn't seem to accurately map to what is expected by the
function.

The result is that upon creation of an issue, we're often told "Sub-Task
is not a symbol". By simplifying the call, this disappears.